refactor: PlaylistObject型をPlaylistType.tsに移動 #126
