+++ src/types/playlistType.ts
@@ -1,0 +1,8 @@
+export type PlaylistObject = {
+  albumImages: string[];
+  createdAt: any;
+  id: string;
+  name: string;
+  totalDuration: number;
+  trackCount: number;
+};
